Add files via upload
# Kickstarting with Excel

## Overview of Project

### Purpose

	'Louisa is producing a play. As someone working in play production, 
Louisa knows that the success of a show is highly dependent on more than how 
good or bad a show is, but also on its funding and when in the year it is 
being produced. Knowing this, Louisa has built a data set using Microsoft Excel
showing not only data on plays, but other "creative projects" (Movies, theater construction,
etc). She has gathered information on fundraising, the time of year fundraisers
were launched, the type of production, where it was produced, etc. Using this data and 
the analysis tools in Microsoft Excel, it will be determined the proper fundraising goal 
and when Louisa should launch her fundraising campaign for her new play "Fever". Louisa
will also gain insight on when to launch a fundraising campaign and how much the fundraising
goal will be for other types of projects as well.

## Analysis and Challenges

### Analysis of Outcomes Based on Launch Date

	'The following data shows the relationship between when a Theater production's 
fundraising is launch vs. the success of that show. The show's outcomes are measured as 
"successful", "failed" or "cancelled". Note that the data gathered is for all theater 
productions, which includes plays but also includes musicals and building projects for 
theaters.
	According to the data, a theater production always has a higher probability of 
being "successful" vs "failed" or "cancelled". However, From January to April there is a
higher chance that a project will fail. The time of the year that projects see most success
kicks off in April, peaks in June and remains high until August. Come August through December,
the chances of success and failure are similar to that of January through April. 
	For a theater production, one should launch their fundraising project sometime between 
April and August for the highest chance of seeing success. 

	This data was collected by creating a pivot chart in Microsoft Excel of the original source data.
The pivot charts filters were the "years" and "Parent Category" columns from the source data. The pivot
table's columns were the "outcomes" (successful, failed or cancelled). The pivot tables Rows was
the date created (measured by months in the year.) And finally, the pivot tables values were labeled
as "count of outcomes". Once this table was created, a line graph was created to present the findings 
visually. 


### Analysis of Outcomes Based on Goals

	'The following data shows the relationship between a play's outcomes and the goal amount for
the fundraising. The play's outcomes are measured by "successful", "Failed", or "Cancelled". The goal 
amounts have been organized into ranges: less than $1000, $1000 to $1499, $1500 to $1999, $2000 to $2499,
continuing up until $45000 to $49999, and ending with Greater than $50000. Once the numbers were found for each 
outcome and goal range, a percentage was found to compare the success and failure of plays depending on
the fundraising goal. 
	According to the data, a plays fundraising campaign has the highest probability of success when the goal is set
anywhere less than or equal to $14999. Once the goal is $15000 or higher, the success is highly unlikely. 
	This data was created using Microsoft Excel's "Countifs" function to pull the data needed from the larger data set.
Countif was used to pull out numbers based on the goal amounts, the outcomes and to specifically only look at these measures
for "plays". Once the numbers were created, a "Sum" function in Microsoft was used to tally up the total amount of projects
that fell in a goal range. From there, percentages were taken based on the outcomes. Finally, Microsoft Excel was
used to create a line graph to display the data visually. 

### Challenges and Difficulties Encountered
	One of the main challenges to this project was the size of the original dataset. Though robust, the size of the original excel 
sheet made way for mistakes to occur. Also, the original dataset was not organized in ways that were the most helpful, so it created
the challenge or reorganizing the data in a way that was useful to the questions at hand. 

## Results

- What are two conclusions you can draw about the Outcomes based on Launch Date?
	1)The amount of theater projects that fail stays constant through the year.
	2)The amount of theater projects that succeed are always higher than those who fail throughout the year, but 
	success peaks when a campaign is launched in June. 

- What can you conclude about the Outcomes based on Goals?
	For a play, it is best to keep your Fundraising goal under $15000 to see your campaign be successful. 

- What are some limitations of this dataset?
	The dataset I've created gives no insight into the genre or story of a play, which could give insight
	into whether a project would become successful. 

- What are some other possible tables and/or graphs that we could create?
	A table could be created to show the success based on the genre or type of story a play is and its outcomes.